Validation system, fixed layout slightly
- Basic validation system checks that the sequence given in terms of a list can be parsed,
that the OEIS code given exists in sloane_functions.py, and that the parameters of for a
given built-in sequence makes sense. For example you can't ignore the coefficient list
when choosing a linear recurrence, and you can't give some text as a mod parameter. Right
now the feedback is given via a small output panel on the right. I'm hoping this is temporary
until someone has a better idea of how to display the errors.

- The layout looks much better on chrome now, but the canvas still overflows slightly
below and to the right.